Changes to allow non-combinatoric distances (Still incomplete support for this)
-Changed distances array to double
-Added input to Context construct to allow for arbitrary distances matrix
-Added input to fromFile, allowing caller to specify that the distances from theembedding should be used
-Added options to resampling.

+++ networkReliabilityCommon/NetworkReliabilityObs.cpp
@@ -77,7 +77,7 @@ namespace networkReliability
 		
 		const EdgeState* oldEdgeStatesPtr = state.get();
 
-		const int* edgeDistances = context.getEdgeDistances();
+		const double* edgeDistances = context.getEdgeDistances();
 
 		std::size_t sourceEdge = 0;
 		while(sourceEdge < nEdges)
